#10 - print out progress bar in synthetic_comparison
