+++ requirements.txt
@@ -2,3 +2,4 @@ numpy == 1.14.2
 scipy == 0.19.1
 scikit-learn == 0.19.1
 pandas == 0.23.1
+tqdm == 4.23.4
